+++ parser/model/BUILD.bazel
@@ -30,9 +30,10 @@ go_test(
     ],
     embed = [":model"],
     flaky = True,
-    shard_count = 20,
+    shard_count = 21,
     deps = [
         "//parser/charset",
+        "//parser/duration",
         "//parser/mysql",
         "//parser/terror",
         "//parser/types",
