+++ parser/model/model.go
@@ -1757,6 +1757,9 @@ func (p *PolicyInfo) Clone() *PolicyInfo {
 // DefaultJobInterval sets the default interval between TTL jobs
 const DefaultJobInterval = time.Hour
 
+// DefaultJobIntervalStr is the string representation of DefaultJobInterval
+const DefaultJobIntervalStr = "1h"
+
 // TTLInfo records the TTL config
 type TTLInfo struct {
 	ColumnName      CIStr  `json:"column"`
